hubby_backend/requirements.txt
Django==4.2.15
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.0
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
pymongo==4.6.1
python-dateutil==2.8.2
dnspython==2.6.1
certifi==2024.2.2
gunicorn==21.2.0
firebase-admin==6.4.0
cryptography==42.0.2
PyJWT==2.12.1